+++ lib/factory_provider.dart
@@ -3,4 +3,5 @@ library;
 
 export 'package:factory_core/factory_core.dart';
 
+export 'src/factory_change_notifier.dart';
 export 'src/factory_scope.dart';